moved unparsed atom handlers to analysis.
